requirements.txt
tts
whisper
openai
sqlite3
faiss-cpu
argostranslate
pydub
requests
flask
twilio
python-dotenv
